requirements.txt
Flask==2.2.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.2